Change event_queue.id back to BIGSERIAL
Prior to this commit, we were "conserving" event IDs by resetting the
sequence to `max(event_queue.id)` after cleaning up completed events. We
originally used BIGSERIAL for this column (hence why
event_failure_reason.event is still BIGINT), but this was changed in
78b20221a809d2fcdab86f619569b591f78f8d0e with the rationale that we'd
never need a SERIAL (let alone BIGSERIAL) number of events in the table
at once.

This rationale no longer makes sense to me, because there's no reason to
want to conserve IDs in this manner to allow a smaller type, other than
saving an inconsequential number of bytes; the table isn't expected to
have too many rows at once to begin with. Resetting the sequence also
feels fragile and confusing.

+++ sql/create.sql
@@ -53,12 +53,12 @@ CREATE TYPE event_state AS ENUM (
 );
 
 CREATE TABLE event_queue (
-    id                  SERIAL,
+    id                  BIGSERIAL,
     state               event_state NOT NULL DEFAULT 'queued',
     entity_type         indexable_entity_type NOT NULL,
     action              event_queue_action NOT NULL,
     message             JSONB NOT NULL,
-    depends_on          INTEGER[],
+    depends_on          BIGINT[],
     created             TIMESTAMP WITH TIME ZONE NOT NULL DEFAULT NOW(),
     -- Note `event_queue_idx_queued_uniq` below. Due to the requirement
     -- that queued events be unique, external triggers should have an
